Test suite updated
- Base class for sessionizer tests
- Test for the type of events processed by
WebActivitySessionizerSketchPlugin
- _create_mock_event changed to accept custom attributes for the event
source
